release 3.1.72 source code

+++ Services/Kafka/V2/Model/GroupInfoSimple.cs
@@ -50,7 +50,7 @@ namespace HuaweiCloud.SDK.Kafka.V2.Model
         /// 堆积数。
         /// </summary>
         [JsonProperty("lag", NullValueHandling = NullValueHandling.Ignore)]
-        public int? Lag { get; set; }
+        public long? Lag { get; set; }
